Added restriction pages
User can no longer edit or delete when not the host. Stop user from login twice.

+++ base/views.py
@@ -5,8 +5,9 @@ from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
-
+from django.http import HttpResponse
 # rooms= [
 #     {'id':1, 'name':'Lets learn Physic!'},
 #     {'id':2, 'name':'Lets learn Math!'},
@@ -16,6 +17,9 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -63,6 +67,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -75,10 +80,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request,'base/room_make.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     # this form will be pre filled
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Nope, you are not allowed here')
 
     if request.method == 'POST' :
         form= RoomForm(request.POST, instance=room)
@@ -89,8 +98,13 @@ def updateRoom(request,pk):
     context = {'form': form}
     return render(request,'base/room_make.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Nope, you are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
